backend/requirements.txt
python-dotenv
flask
flask-cors
langchain
langchain-openai
langchain-community
langchain_chroma
langchain_core
langchain_text_splitters
requests
bs4
pymongo
flask-bcrypt